typescript d.ts inferred from classes hierarchy TypeInfo->Properties and some public interfaces from the Jsonix.Context
