+++ .gitignore
@@ -7,4 +7,5 @@ target
 .idea
 /bin/
 *.iml
+.classpath
 rm_ide_settings.sh
