+++ typescript/src/main/java/org/hisrc/jsonix/Model/AbstractElementsPropertyInfo.java
@@ -1,7 +1,9 @@
 package org.hisrc.jsonix.Model;
 
+import org.hisrc.jsonix.Schema.XML.QName;
+
 public class AbstractElementsPropertyInfo extends PropertyInfo implements org.hisrc.jsonix.Binding.Mashalls.Element, org.hisrc.jsonix.Binding.Unmashalls.Element{
-    Object wrapperElementName = null;
+    QName wrapperElementName = null;
     boolean allowDom = false;
     boolean allowTypedObject = true;
     boolean mixed = false;
