+++ typescript/src/main/java/org/hisrc/jsonix/Model/ElementPropertyInfo.java
@@ -1,7 +1,10 @@
 package org.hisrc.jsonix.Model;
 
+import org.hisrc.jsonix.Schema.XML.QName;
+
 public class ElementPropertyInfo extends AbstractElementsPropertyInfo implements org.hisrc.jsonix.Binding.Mashalls.Element{
-    String typeInfo = "String";
-    Object elementName = null;
+    //May be a String
+    TypeInfo typeInfo = null;
+    QName elementName = null;
 
 }
