+++ typescript/src/main/java/org/hisrc/jsonix/Model/TypeInfo.java
@@ -1,6 +1,6 @@
 package org.hisrc.jsonix.Model;
 
 public class TypeInfo {
-    Object name = null;
-    Object baseTypeInfo = null;
+    String name = "";
+    TypeInfo baseTypeInfo = null;
 }
