+++ typescript/src/main/java/org/hisrc/jsonix/Schema/XML/QName.java
@@ -1,0 +1,11 @@
+package org.hisrc.jsonix.Schema.XML;
+
+public class QName 
+{
+	String key = null;
+	String 	namespaceURI = null;
+	String 	localPart = null;
+	String 	prefix = null;
+	String 	string = null;
+
+}
