--- typescript/src/main/java/org/hisrc/jsonix/Schema/XSD/AnySimpleType.java
@@ -1,8 +1,0 @@
-package org.hisrc.jsonix.Schema.XSD;
-
-import org.hisrc.jsonix.Model.TypeInfo;
-
-public class AnySimpleType extends TypeInfo
-{
-
-}
