--- typescript/src/main/java/org/hisrc/jsonix/Schema/XSD/QName.java
@@ -1,6 +1,0 @@
-package org.hisrc.jsonix.Schema.XSD;
-
-public class QName extends AnySimpleType
-{
-	String name = "Qname";
-}
